multi: move channel announcement validation to netann
from the graph package.
